+++ src/main/java/org/folio/sidecar/integration/cred/model/UserCredentials.java
@@ -1,10 +1,12 @@
 package org.folio.sidecar.integration.cred.model;
 
+import io.quarkus.runtime.annotations.RegisterForReflection;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.ToString;
 
 @Data
+@RegisterForReflection
 @AllArgsConstructor(staticName = "of")
 public class UserCredentials {