+++ src/main/java/org/folio/sidecar/integration/keycloak/model/TokenIntrospectionResponse.java
@@ -1,9 +1,11 @@
 package org.folio.sidecar.integration.keycloak.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import io.quarkus.runtime.annotations.RegisterForReflection;
 import lombok.Data;
 
 @Data
+@RegisterForReflection
 public class TokenIntrospectionResponse {
 
   public static final TokenIntrospectionResponse INACTIVE_TOKEN = new TokenIntrospectionResponse();