+++ src/main/java/org/folio/sidecar/model/EntitlementsEvent.java
@@ -1,5 +1,6 @@
 package org.folio.sidecar.model;
 
+import io.quarkus.runtime.annotations.RegisterForReflection;
 import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,6 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor(staticName = "of")
+@RegisterForReflection
 public class EntitlementsEvent {
 
   public static final String ENTITLEMENTS_EVENT = "entitlements";